+++ CHANGELOG.md
@@ -11,7 +11,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- [JUnit Platform Engine] Add constant for fixed.max-pool-size property ([#2713](https://github.com/cucumber/cucumber-jvm/pull/2713) M.P. Korstanje)
 - [Core] Support directories containing exclusively rerun files using the `@path/to/rerun` syntax ([#2710](https://github.com/cucumber/cucumber-jvm/pull/2710) Daniel Whitney, M.P. Korstanje)
+
+## [7.11.2] - 2023-03-23
+### Fixed
+- [JUnit Platform Engine] Corrupted junit-xml report when using `surefire.rerunFailingTestsCount` parameter ([#2709](https://github.com/cucumber/cucumber-jvm/pull/2709) M.P. Korstanje)
 
 ## [7.11.1] - 2023-01-27
 ### Added
@@ -227,7 +232,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [Weld] Removed `cucumber-weld` in favour of `cucumber-jakarta-cdi` or `cucumber-cdi2`. ([#2276](https://github.com/cucumber/cucumber-jvm/issues/2276) M.P. Korstanje)
 - [Needle] Removed `cucumber-needled` in favour of `cucumber-jakarta-cdi` or `cucumber-cdi2`. ([#2276](https://github.com/cucumber/cucumber-jvm/issues/2276) M.P. Korstanje)
 
-[Unreleased]: https://github.com/cucumber/cucumber-jvm/compare/v7.11.1...main
+[Unreleased]: https://github.com/cucumber/cucumber-jvm/compare/v7.11.2...main
+[7.11.2]: https://github.com/cucumber/cucumber-jvm/compare/v7.11.1...main
 [7.11.1]: https://github.com/cucumber/cucumber-jvm/compare/v7.11.0...main
 [7.11.0]: https://github.com/cucumber/cucumber-jvm/compare/v7.10.1...v7.11.0
 [7.10.1]: https://github.com/cucumber/cucumber-jvm/compare/v7.10.0...v7.10.1